[Packaging] Review rubygem package
Let's include the gems as sources instead of build requirements so
building this package does not require so many gem packages.

Build requiring rubygem packages is a "nightmare" when you try to build
portus in different repos:

* sometimes the rubygem RPMs are not there and need to be submited
* othertimes they are present but have a different version and that
 creates a conflict

Since the new macros in ruby-common are generating the

"Provides: bundled(rubygem(foo)) = version"

we don't need the BR to explicetely tell which rubygems are in, plus the
"provides" is much more explicit.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>